# API-Spec/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PaymentGateway API",
    "version": "v1"
  },
  "paths": {
    "/Payments/{paymentId}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Retrieve the details of a previously made payment",
        "parameters": [
          {
            "name": "PaymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The payment was found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/Payments/create": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Process a payment through the payment gateway and receive either a successful or unsuccessful response",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequestModel"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The payment was created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePaymentResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePaymentResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePaymentResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PaymentStatusCode": {
        "enum": [
          "PaymentSuccess",
          "PaymentFailure"
        ],
        "type": "string"
      },
      "CreditCardModel": {
        "type": "object",
        "properties": {
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "expiryDateMonth": {
            "type": "integer",
            "format": "int32"
          },
          "expiryDateYear": {
            "type": "integer",
            "format": "int32"
          },
          "ccv": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PaymentModel": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string",
            "format": "uuid"
          },
          "merchantId": {
            "type": "string",
            "format": "uuid"
          },
          "externalShopperIdentifier": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatusCode"
              }
            ]
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currencyCode": {
            "type": "string",
            "nullable": true
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "creditCard": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CreditCardModel"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetPaymentResponseModel": {
        "type": "object",
        "properties": {
          "payment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentModel"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResponseModel": {
        "type": "object",
        "properties": {
          "errorDescription": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "CreatePaymentRequestModel": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currencyCode": {
            "type": "string",
            "nullable": true
          },
          "externalShopperIdentifier": {
            "type": "string",
            "nullable": true
          },
          "creditCardNumber": {
            "type": "string",
            "nullable": true
          },
          "creditCardExpiryMonth": {
            "type": "integer",
            "format": "int32"
          },
          "creditCardExpiryYear": {
            "type": "integer",
            "format": "int32"
          },
          "creditCardCcv": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreatePaymentResponseModel": {
        "type": "object",
        "properties": {
          "wasPaymentSuccessful": {
            "type": "boolean"
          },
          "paymentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      }
    }
  }
}